requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
supabase==2.28.3
groq==0.13.0
google-genai>=1.0,<2.0
python-docx==1.1.2
feedparser==6.0.11
httpx==0.27.2
python-dateutil==2.9.0
python-telegram-bot==21.9
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
